backend/requirements.txt
Flask
Gunicorn
psycopg2-binary==2.9.5
flask-swagger